hide the LocationIndicatorLayer impl
