+++ MapboxGLAndroidSDK/src/cpp/style/layers/layer.cpp.ejs
@@ -2,7 +2,7 @@
   const type = locals.type;
   const properties = locals.properties;
 -%>
-// This file is generated. Edit android/platform/scripts/generate-style-code.js, then run `make android-style-code`.
+// This file is generated. Edit scripts/generate-style-code.js, then run `make style-code`.
 
 #include "<%- type.replace('-', '_') %>_layer.hpp"
