+++ MapboxGLAndroidSDK/src/cpp/style/layers/location_indicator_layer.cpp
@@ -1,4 +1,4 @@
-// This file is generated. Edit android/platform/scripts/generate-style-code.js, then run `make android-style-code`.
+// This file is generated. Edit scripts/generate-style-code.js, then run `make style-code`.
 
 #include "location_indicator_layer.hpp"
