+++ MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/LayoutPropertyValue.java
@@ -2,9 +2,9 @@ package com.mapbox.mapboxsdk.style.layers;
 
 import androidx.annotation.NonNull;
 
-class LayoutPropertyValue<T> extends PropertyValue<T> {
+public class LayoutPropertyValue<T> extends PropertyValue<T> {
 
-  LayoutPropertyValue(@NonNull String name, T value) {
+  public LayoutPropertyValue(@NonNull String name, T value) {
     super(name, value);
   }
