+++ MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/PaintPropertyValue.java
@@ -2,9 +2,9 @@ package com.mapbox.mapboxsdk.style.layers;
 
 import androidx.annotation.NonNull;
 
-class PaintPropertyValue<T> extends PropertyValue<T> {
+public class PaintPropertyValue<T> extends PropertyValue<T> {
 
-  PaintPropertyValue(@NonNull String name, T value) {
+  public PaintPropertyValue(@NonNull String name, T value) {
     super(name, value);
   }
