+++ MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/layers/property_factory.java.ejs
@@ -1,23 +1,38 @@
 <%
   const paintProperties = locals.paintProperties;
   const layoutProperties = locals.layoutProperties;
+  const locationIndicator = locals.locationIndicator;
 -%>
-// This file is generated. Edit android/platform/scripts/generate-style-code.js, then run `make android-style-code`.
+// This file is generated. Edit scripts/generate-style-code.js, then run `make style-code`.
 
+<% if (locationIndicator) { -%>
+package com.mapbox.mapboxsdk.location;
+<% } else { -%>
 package com.mapbox.mapboxsdk.style.layers;
+<% } -%>
 
 import androidx.annotation.ColorInt;
 
 import com.mapbox.mapboxsdk.style.expressions.Expression;
 import com.mapbox.mapboxsdk.style.types.Formatted;
 import static com.mapbox.mapboxsdk.utils.ColorUtils.colorToRgbaString;
+<% if (locationIndicator) { -%>
+import com.mapbox.mapboxsdk.style.layers.LayoutPropertyValue;
+import com.mapbox.mapboxsdk.style.layers.PaintPropertyValue;
+import com.mapbox.mapboxsdk.style.layers.Property;
+import com.mapbox.mapboxsdk.style.layers.PropertyValue;
+<% } -%>
 
 /**
  * Constructs paint/layout properties for Layers
  *
  * @see <a href="https://www.mapbox.com/mapbox-gl-style-spec/#layers">Layer style documentation</a>
  */
+<% if (locationIndicator) { -%>
+class LocationPropertyFactory {
+<% } else { -%>
 public class PropertyFactory {
+<% } -%>
 
   /**
    * Set the property visibility.
