+++ MapboxGLAndroidSDKTestApp/src/androidTest/java/com/mapbox/mapboxsdk/testapp/style/layer.junit.ejs
@@ -2,9 +2,13 @@
   const type = locals.type;
   const properties = locals.properties;
 -%>
-// This file is generated. Edit android/platform/scripts/generate-style-code.js, then run `make android-style-code`.
+// This file is generated. Edit scripts/generate-style-code.js, then run `make style-code`.
 
+<% if (type === 'location-indicator') { -%>
+package com.mapbox.mapboxsdk.location;
+<% } else { -%>
 package com.mapbox.mapboxsdk.testapp.style;
+<% } -%>
 
 import android.graphics.Color;
 import androidx.test.annotation.UiThreadTest;
@@ -15,7 +19,9 @@ import org.junit.Before;
 import timber.log.Timber;
 
 import com.mapbox.mapboxsdk.style.expressions.Expression;
+<% if (type !== 'location-indicator') { -%>
 import com.mapbox.mapboxsdk.style.layers.<%- camelize(type) %>Layer;
+<% } -%>
 <% if (type === 'symbol') { -%>
 import com.mapbox.mapboxsdk.style.types.Formatted;
 import com.mapbox.mapboxsdk.style.types.FormattedSection;
@@ -27,7 +33,11 @@ import org.junit.runner.RunWith;
 import static com.mapbox.mapboxsdk.style.expressions.Expression.*;
 import static org.junit.Assert.*;
 import static com.mapbox.mapboxsdk.style.layers.Property.*;
+<% if (type === 'location-indicator') { -%>
+import static com.mapbox.mapboxsdk.location.LocationPropertyFactory.*;
+<% } else { -%>
 import static com.mapbox.mapboxsdk.style.layers.PropertyFactory.*;
+<% } -%>
 
 import com.mapbox.mapboxsdk.style.layers.TransitionOptions;
